refactor: 엔티티 setter, builder 에노테이션 제거 및 팩토리 메서드 추가, 유저 유저 프로필 서비스 로직 변경
1. 모든 엔티티 setter, builder 에토테이션 제거, 팩토리 메서드 추가
2. 유저, 유저 프로필 엔티티 수정 반영 + 예외처리 코드 변경

+++ src/main/java/com/hellofit/hellofit_server/user/profile/dto/UpdateUserProfileRequestDto.java
@@ -3,6 +3,8 @@ package com.hellofit.hellofit_server.user.profile.dto;
 import com.hellofit.hellofit_server.user.profile.UserProfileEntity;
 import jakarta.validation.constraints.*;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -13,9 +15,22 @@ import java.util.List;
 public class UpdateUserProfileRequestDto {
     UserProfileEntity.AgeGroup ageGroup;
     UserProfileEntity.Gender gender;
-    @DecimalMin("50.0") @DecimalMax("300.0") Double height;
-    @DecimalMin("20.0") @DecimalMax("500.0")  Double weight; // kg
-    @Min(0) @Max(24 * 60) Integer sleepMinutes;
-    @Min(0) @Max(24 * 60) Integer exerciseMinutes;
+
+    @DecimalMin("50.0")
+    @DecimalMax("300.0")
+    BigDecimal height;
+
+    @DecimalMin("20.0")
+    @DecimalMax("500.0")
+    BigDecimal weight; // kg
+
+    @Min(0)
+    @Max(24 * 60)
+    Integer sleepMinutes;
+
+    @Min(0)
+    @Max(24 * 60)
+    Integer exerciseMinutes;
+    
     List<@NotBlank @Size(max = 100) String> forbiddenFoods;
 }
